ref : consumer 컨트롤러 구조 통합
