--- homecare/src/main/java/jaega/homecare/domain/serviceRequest/controller/ServiceRequestControllerImpl.java
@@ -1,77 +1,0 @@
-package jaega.homecare.domain.serviceRequest.controller;
-
-import jaega.homecare.domain.serviceRequest.dto.req.ConsumerServiceRequest;
-import jaega.homecare.domain.serviceRequest.dto.req.UpdateServiceRequest;
-import jaega.homecare.domain.serviceRequest.dto.res.GetCreateServiceResponse;
-import jaega.homecare.domain.serviceRequest.dto.res.GetServiceRequestById;
-import jaega.homecare.domain.serviceRequest.dto.res.GetServiceRequestResponse;
-import jaega.homecare.domain.serviceRequest.entity.ServiceRequest;
-import jaega.homecare.domain.serviceRequest.entity.ServiceRequestStatus;
-import jaega.homecare.domain.serviceRequest.service.command.ServiceRequestCommandService;
-import jaega.homecare.domain.serviceRequest.service.query.ServiceRequestQueryService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-import java.util.UUID;
-
-@RequiredArgsConstructor
-@RestController
-@RequestMapping("/api/consumer/request")
-public class ServiceRequestControllerImpl implements ServiceRequestController{
-    private final ServiceRequestCommandService serviceRequestCommandService;
-    private final ServiceRequestQueryService serviceRequestQueryService;
-
-    /**
-     *
-     * 수요자 서비스 요청 API
-     */
-    @Override
-    public ResponseEntity<GetCreateServiceResponse> createServiceRequest(@RequestBody ConsumerServiceRequest request){
-        GetCreateServiceResponse response = serviceRequestCommandService.createServiceRequest(request);
-        return ResponseEntity.ok(response);
-    }
-
-    /**
-     *
-     * 수요자가 신청한 서비스 내역 조회 API
-     */
-    @Override
-    public ResponseEntity<List<GetServiceRequestResponse>> getConsumerServiceRequest(@RequestParam UUID consumerId){
-        List<GetServiceRequestResponse> response = serviceRequestQueryService.findConsumerRequests(consumerId);
-        return ResponseEntity.ok(response);
-    }
-
-    /**
-     *
-     * 수요자가 신청한 서비스 내역 조회 API (신청 서비스 상태 조건)
-     */
-    @Override
-    public ResponseEntity<List<GetServiceRequestResponse>> getConsumerServiceRequestByStatus(@RequestParam UUID consumerId, ServiceRequestStatus status){
-        List<GetServiceRequestResponse> response = serviceRequestQueryService.findConsumerRequestsByStatus(consumerId, status);
-        return ResponseEntity.ok(response);
-    }
-
-    /**
-     *
-     * 수요자가 신청한 서비스 내역 상세 조회
-     */
-    @Override
-    public ResponseEntity<GetServiceRequestById> getServiceRequestById(UUID serviceRequestId) {
-        GetServiceRequestById response = serviceRequestQueryService.findServiceRequestById(serviceRequestId);
-        return ResponseEntity.ok(response);
-    }
-
-    @Override
-    public ResponseEntity<Void> updateServiceRequest(@RequestParam UUID serviceRequestId,
-                                                     @RequestBody UpdateServiceRequest request){
-        ServiceRequest serviceRequest = serviceRequestQueryService.getServiceRequest(serviceRequestId);
-        serviceRequestCommandService.updateServiceRequest(request, serviceRequest);
-        return ResponseEntity.noContent().build();
-    }
-
-}
